+++ python/bovnar/enums.py
@@ -455,6 +455,7 @@ class BaseUnit(IntEnum):
     CIRCULAR_MIL       = 100258   # cml
     UNIT_POLE          = 100259   # unit_pole
     HORSEPOWER_WATER   = 100260   # hp_W
+    MONTH_GREGORIAN    = 100261   # mo_greg
 
     # ---- Currencies: block 90 -------------------------------------------
     #
@@ -696,7 +697,7 @@ class BaseUnit(IntEnum):
 UNIT_BLOCK_SIZE     = 10000
 
 UNIT_NATIVE_FIRST   = BaseUnit.BIT
-UNIT_NATIVE_LAST    = BaseUnit.HORSEPOWER_WATER
+UNIT_NATIVE_LAST    = BaseUnit.MONTH_GREGORIAN
 
 CURRENCY_FIRST      = BaseUnit.AED
 CURRENCY_LAST       = BaseUnit.RUNE